requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
python-dotenv==1.0.1
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.9.0
Werkzeug==3.0.1
